Publisher: Extract all duplicated methods to a base class
Was trying to fix #71, and noticed we had to do the same thing in
two places, so I extracted the methods to a base class instead.

I don't remember if there was the reason to why we have two similar
classes though.
