+++ lib/twingly/amqp.rb
@@ -4,6 +4,7 @@ require "twingly/amqp/connection"
 require "twingly/amqp/subscription"
 require "twingly/amqp/ping_options"
 require "twingly/amqp/pinger"
+require "twingly/amqp/base_publisher"
 require "twingly/amqp/default_exchange_publisher"
 require "twingly/amqp/topic_exchange_publisher"
 require "twingly/amqp/null_logger"
