+++ lib/twingly/amqp/default_exchange_publisher.rb
@@ -1,47 +1,14 @@
-require "twingly/amqp/connection"
-require "json"
-require "ostruct"
+require "twingly/amqp/base_publisher"
 
 module Twingly
   module AMQP
-    class DefaultExchangePublisher
+    class DefaultExchangePublisher < BasePublisher
       def initialize(queue_name:, connection: nil)
+        super(connection: connection)
+
         options.routing_key = queue_name
 
-        connection ||= Connection.instance
-        @exchange = connection.create_channel.default_exchange
-      end
-
-      def publish(message)
-        raise ArgumentError unless message.respond_to?(:to_h)
-
-        payload = message.to_h.to_json
-        opts    = options.to_h
-        @exchange.publish(payload, opts)
-      end
-
-      # only used by tests to avoid sleeping
-      def publish_with_confirm(message)
-        channel = @exchange.channel
-        channel.confirm_select unless channel.using_publisher_confirmations?
-
-        publish(message)
-
-        @exchange.wait_for_confirms
-      end
-
-      def configure_publish_options
-        yield options
-      end
-
-      private
-
-      def options
-        @options ||=
-          OpenStruct.new(
-            content_type: "application/json",
-            persistent: true,
-          )
+        @exchange = @connection.create_channel.default_exchange
       end
     end
   end
